Split the stack trace tests out of `JoeTest`.
